[docs-no-version] 2.2/ only keep necessary useVersion and remove all others (#15779)
## Summary & Motivation
depends on https://github.com/dagster-io/dagster/pull/15777

this completes the `useVersion` clean up.
- it keeps necessary version info in useVersion: `version` shows the latest version, `showVersionBanner` indicates whether or not to show the version notice banner
  * note: this is useful in previews because going forward we're going to rely on previews to show docs of older version. when a user lands on a older version (preview), we'd always show the warning banner to remind them that the site is an outdated site.
- it moves path info to a separate `usePath` file/function. previously, because version has been part of the path, we derive the normalized path from useVersion. now, as we only render latest version (master), we no longer need that derived logic.
  * this also results in removing useVersion usage in a few places: Link, Sidebar, Pagination (we aren't actively using it anyways)

Next:
- update release pipeline
- more cleanups such as func names

## How I Tested These Changes
click around in preview: internal links and sidebar links all work
