+++ docs/next/components/Sidebar.tsx
@@ -5,13 +5,11 @@ import NextLink from 'next/link';
 import React, {useState} from 'react';
 
 import Icons from '../components/Icons';
-import {useNavigation, flatten, getNavKey, getNavLvl} from '../util/useNavigation';
-import {useVersion} from '../util/useVersion';
-
-import Link from './Link';
+import navigation, {flatten, getNavKey, getNavLvl} from '../util/navigation';
+import {usePath} from '../util/usePath';
 
 const useCurrentSection = (navigation) => {
-  const {asPath} = useVersion();
+  const {asPath} = usePath();
   const match = navigation.find((item) => item.path !== '/' && asPath.startsWith(item.path));
   return match || navigation.find((item) => item.path === '/');
 };
@@ -108,19 +106,10 @@ const MenuItem = React.forwardRef<HTMLAnchorElement, React.PropsWithChildren<Men
       return linkElement;
     }
 
-    if (item.isUnversioned) {
-      return (
-        <NextLink href={item.path} passHref>
-          {linkElement}
-        </NextLink>
-      );
-    }
-
-    // Versioned link
     return (
-      <Link href={item.path} passHref>
+      <NextLink href={item.path} passHref>
         {linkElement}
-      </Link>
+      </NextLink>
     );
   },
 );
@@ -187,8 +176,7 @@ const RecursiveNavigation = ({
   navKeysToExpanded,
   setNavKeysToExpanded,
 }) => {
-  const {asPathWithoutAnchor} = useVersion();
-  const navigation = useNavigation();
+  const {asPathWithoutAnchor} = usePath();
   const currentSection = useCurrentSection(navigation);
   const navKey = getNavKey(parentKey, idx);
   const lvl = getNavLvl(navKey);
@@ -258,8 +246,6 @@ const RecursiveNavigation = ({
 };
 
 const TopLevelNavigation = () => {
-  const navigation = useNavigation();
-
   const map = {};
   const [navKeysToExpanded, setNavKeysToExpanded] = useState<{
     [key: string]: boolean;
