+++ docs/next/pages/sitemap.xml.tsx
@@ -1,4 +1,4 @@
-import {latestAllPaths} from 'util/useNavigation';
+import {latestAllPaths} from 'util/navigation';
 
 const toUrl = (host, route) => `<url><loc>http://${host}${route}</loc></url>`;
